pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentsuite"
version = "1.0.7"
description = "Seven role-specific reasoning agents that turn vague intent into precise operating artifacts"
authors = [{name = "Scott Converse"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["ai", "agents", "mcp", "anthropic", "openai", "specification", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pydantic>=2.5,<3",
    "tenacity>=8.2,<10",
    "typer>=0.12,<1",
    "httpx>=0.27,<1",
    "jinja2>=3.1,<4",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40,<1"]
openai = ["openai>=1.50,<3"]
gemini = ["google-genai>=1.0,<2"]
ollama = ["ollama>=0.4,<1"]
mcp = ["mcp>=1.0,<2"]
image = ["pillow>=10.0,<13"]
all = ["agentsuite[anthropic,openai,gemini,ollama,mcp,image]"]
ocr = ["pytesseract>=0.3"]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.6",
    "mypy>=1.10",
    "vcrpy>=6.0",
    "build>=1.2",
    "twine>=5.0",
    "agentsuite[anthropic,openai,gemini,ollama,mcp,image]",
]

[project.scripts]
agentsuite = "agentsuite.cli:app"
agentsuite-mcp = "agentsuite.mcp_server:main"

[project.urls]
Repository = "https://github.com/scottconverse/AgentSuite"
Issues = "https://github.com/scottconverse/AgentSuite/issues"
Documentation = "https://scottconverse.github.io/AgentSuite/"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentsuite*"]

[tool.setuptools.package-data]
"agentsuite" = ["py.typed"]
"agentsuite.agents.founder" = ["templates/*.md", "prompts/*.jinja2"]
"agentsuite.agents.design" = ["prompts/*.jinja2", "templates/*.md"]
"agentsuite.agents.engineering" = ["prompts/*.jinja2", "templates/*.md"]
"agentsuite.agents.marketing" = ["prompts/*.jinja2", "templates/*.md"]
"agentsuite.agents.trust_risk" = ["prompts/*.jinja2", "templates/*.md"]
"agentsuite.kernel" = ["prompts/*.jinja2"]
"agentsuite.agents.product" = ["prompts/*.jinja2", "templates/*.md"]
"agentsuite.agents.cio" = ["prompts/*.jinja2", "templates/*.md"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-m 'not cleanroom and not live and not live_ollama'"
markers = [
    "live: requires live LLM API access (gated by RUN_LIVE_TESTS=1)",
    "live_ollama: requires running Ollama daemon (gated by RUN_LIVE_OLLAMA_TESTS=1; zero cost)",
    "slow: slow integration tests",
    "cleanroom: full cleanroom E2E (gated; invoke with `pytest -m cleanroom`)",
]
